# contract/openapi.yaml
openapi: 3.0.3
info:
  title: BuryMe API
  version: 1.0.0
  description: |
    Backend contract for **BuryMe** — a peer-to-peer IOU / micro-lending tracker.
    Derived directly from PRD v2.7, §5 (Conceptual Model), §6–7 (Requirements &
    Scenarios), §8 (Entity Attribute Reference), §11 (Dependencies & Integrations),
    and §12 (Validation & Error Handling).

    ## Architecture
    - **Thin clients.** All business logic, persistence, and third-party integration
      live in the Express/Node backend (§11.0). React (web) and Flutter (mobile)
      consume this API only — neither has direct DB or third-party access.
    - **Auth is Firebase-backed** (§11.1.2). Phone/email OTP and credential
      management happen entirely client-side via the Firebase Auth SDK. The client
      sends the resulting Firebase ID token as `Authorization: Bearer <token>` on
      every request; Express verifies it via the Firebase Admin SDK. There is
      therefore no `/login`, `/send-otp`, or `/verify-otp` endpoint here — those
      never touch this API.
    - **First login provisions a BuryMe user.** A valid Firebase token with no
      matching BuryMe profile yields `404` on `GET /auth/me`; the client then calls
      `POST /auth/register`.

    ## Request vs. Obligation — the core modeling decision
    Per §5 and §8.5/§8.6, **Request** and **Obligation** are two distinct persisted
    resources, not two states of one resource:
...
      It never re-enters a negotiation state.

...
    pre-filled.
...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /settlements/suggestions:
    get:
      tags: [Settlements]
      summary: List settlement suggestions involving the current user
      operationId: listSettlementSuggestions
      responses:
        "200":
          description: Suggestions involving the current user.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SettlementSuggestion" }
    post:
      tags: [Settlements]
      summary: Suggest settling two reciprocal obligations against each other
      description: >
        Either party may initiate once both obligations are open — whoever
        acts first is the proposer, and their own response is recorded as
        `Accepted` immediately, leaving the other party `Pending` (§8.14).
        The server pairs the obligations itself: it requires exactly one
        open obligation in each direction between the two users, and
        `user_a` is always the initiator, so `obligation_id_a` is the
        obligation the initiator borrows on. The net is the difference
        between the two outstanding balances, owed by whoever owes more.
        Nothing settles until both parties accept.
      operationId: createSettlementSuggestion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSettlementSuggestionInput" }
      responses:
        "201":
          description: Suggestion created, awaiting the other party's response.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettlementSuggestion" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "409":
          description: >
            `STATUS_CONFLICT` when there is no eligible reciprocal pair
            between the two users, or more than one open obligation in
            either direction; `DUPLICATE_SUBMISSION` when a pending
            suggestion already exists for this pair.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /settlements/suggestions/{suggestionId}:
    get:
      tags: [Settlements]
      summary: Get a settlement suggestion
      operationId: getSettlementSuggestion
      parameters:
        - $ref: "#/components/parameters/SuggestionId"
      responses:
        "200":
          description: Suggestion detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettlementSuggestion" }
        "404": { $ref: "#/components/responses/NotFound" }
  /settlements/suggestions/{suggestionId}/accept:
    post:
      tags: [Settlements]
      summary: Accept a settlement suggestion
      description: >
        Records this user's response. The suggestion's aggregate `status` becomes
        `Accepted` — and both obligations move to `Settled` — only once both
        `user_a_response` and `user_b_response` are `Accepted` (§8.14).
      operationId: acceptSettlementSuggestion
      parameters:
        - $ref: "#/components/parameters/SuggestionId"
      responses:
        "200":
          description: Updated suggestion (still `Pending` if awaiting the other party, or `Accepted`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettlementSuggestion" }
        "409":
          description: One of the paired obligations was already settled by other means.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /settlements/suggestions/{suggestionId}/decline:
    post:
      tags: [Settlements]
      summary: Decline a settlement suggestion
      description: Either party declining moves the aggregate `status` to `Declined`; obligations remain unchanged.
      operationId: declineSettlementSuggestion
      parameters:
        - $ref: "#/components/parameters/SuggestionId"
      responses:
        "200":
          description: Suggestion `Declined`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettlementSuggestion" }

  # ══════════════════════════════════════════════════════════════
  # HISTORY (§6.5.1, §8.10) — immutable, read-only
  # ══════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Firebase ID token (1-hour expiry, §11.1.2). Verified on the backend via the
        Firebase Admin SDK on every protected route.
  parameters:
...
    SuggestionId:
      name: suggestionId
      in: path
      required: true
      schema: { type: string }
  responses:
    ValidationError:
      description: One or more input fields failed validation (§12.3.2 `VALIDATION_ERROR`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: VALIDATION_ERROR
              message: The due date must be a future date.
              field: proposed_due_date
...
    NotFound:
      description: Record does not exist or is not accessible to this user (§12.3.2 `NOT_FOUND`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    ErrorCode:
      type: string
      description: Standard error codes per §12.3.2.
      enum:
        - VALIDATION_ERROR
        - DUPLICATE_SUBMISSION
        - STATUS_CONFLICT
        - BALANCE_EXCEEDED
        - CONCURRENT_MODIFICATION
        - PAYMENT_GATEWAY_UNAVAILABLE
        - PAYMENT_GATEWAY_ERROR
        - RECIPIENT_UNVERIFIED
        - PARTIAL_WRITE_ROLLED_BACK
        - UNAUTHENTICATED
        - UNAUTHORIZED
        - NOT_FOUND
        - INTERNAL_ERROR
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Plain-language, user-safe. Never exposes stack traces, DB ids, or raw third-party payloads (§12.1).
            field:
              type: string
              nullable: true
              description: Present only for field-level validation errors; omitted for system errors (§12.3.1).

    # ── Enums (verbatim from §8) ─────────────────────────────────
...
    SettlementResponse:
      type: string
      enum: [Pending, Accepted, Declined]
...
    Money:
      type: object
      description: All amounts in ETB, ≤2 decimal places (§12.1).
      required: [amount, currency]
      properties:
        amount: { type: number, format: double, minimum: 0.01, maximum: 1000000, multipleOf: 0.01 }
        currency: { type: string, default: ETB, enum: [ETB] }

    # ── Auth / Users ──────────────────────────────────────────────
...
    UserSummary:
      type: object
      required: [user_id, display_name, profile_photo_url]
      properties:
        user_id: { type: string }
        display_name: { type: string }
        profile_photo_url: { type: string, nullable: true }
...
    SettlementSuggestion:
      type: object
      required:
        [
          suggestion_id,
          user_a,
          user_b,
          obligation_id_a,
          obligation_id_b,
          net_amount,
          net_payer,
          net_recipient,
          user_a_response,
          user_b_response,
          status,
          created_at,
          resolved_at,
        ]
      properties:
        suggestion_id: { type: string }
        user_a: { $ref: "#/components/schemas/UserSummary" }
        user_b: { $ref: "#/components/schemas/UserSummary" }
        obligation_id_a:
          { type: string, description: Active obligation where user_a is the borrower. }
        obligation_id_b:
          { type: string, description: Active obligation where user_b is the borrower. }
        net_amount: { $ref: "#/components/schemas/Money" }
        net_payer: { $ref: "#/components/schemas/UserSummary" }
        net_recipient: { $ref: "#/components/schemas/UserSummary" }
        user_a_response: { $ref: "#/components/schemas/SettlementResponse" }
        user_b_response: { $ref: "#/components/schemas/SettlementResponse" }
        status: { $ref: "#/components/schemas/SettlementResponse" }
        created_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
    CreateSettlementSuggestionInput:
      type: object
      required: [counterparty_user_id]
      properties:
        counterparty_user_id:
          type: string
          description: >
            The other party. The server finds the reciprocal pair of open
            obligations between the two users itself — the client never
            names the obligations, so it cannot submit a mismatched pair.

    # ── Obligation History (§8.10) ─────────────────────────────────
...